+++ Backend/src/main/java/com/homiwood/peliculas/service/UsuarioService.java
@@ -1,10 +1,11 @@
 package com.homiwood.peliculas.service;
 
+import com.homiwood.peliculas.dto.CrearUsuarioRequest;
+import com.homiwood.peliculas.exception.DuplicateResourceException;
+import com.homiwood.peliculas.exception.NotFoundException;
 import com.homiwood.peliculas.model.Usuario;
 import com.homiwood.peliculas.repository.UsuarioRepository;
 import org.springframework.stereotype.Service;
-import com.homiwood.peliculas.exception.DuplicateResourceException;
-import com.homiwood.peliculas.exception.NotFoundException;
 
 import java.util.List;
 
@@ -21,15 +22,23 @@ public class UsuarioService {
         return usuarioRepository.findAll();
     }
 
-    public Usuario crearUsuario(Usuario usuario) {
+    public Usuario crearUsuario(CrearUsuarioRequest request) {
 
-        if (usuarioRepository.existsByUsername(usuario.getUsername())) {
+        if (usuarioRepository.existsByUsername(request.getUsername())) {
             throw new DuplicateResourceException("El username ya está registrado");
         }
 
-        if (usuarioRepository.existsByEmail(usuario.getEmail())) {
+        if (usuarioRepository.existsByEmail(request.getEmail())) {
             throw new DuplicateResourceException("El email ya está registrado");
         }
+
+        Usuario usuario = new Usuario();
+        usuario.setNombre(request.getNombre());
+        usuario.setUsername(request.getUsername());
+        usuario.setEmail(request.getEmail());
+
+        // Temporal. Más adelante esto se encriptará con Spring Security.
+        usuario.setPasswordHash(request.getPassword());
 
         return usuarioRepository.save(usuario);
     }
@@ -38,7 +47,13 @@ public class UsuarioService {
         return usuarioRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Usuario no encontrado"));
     }
+
     public void eliminarUsuario(Long id) {
+
+        if (!usuarioRepository.existsById(id)) {
+            throw new NotFoundException("Usuario no encontrado");
+        }
+
         usuarioRepository.deleteById(id);
     }
 }
